optional transparent buffer

+++ include/maps.h
@@ -2,6 +2,6 @@
 #define MAPS_H
 #include "defines.h"
 #include "typedefs.h"
-extern const cell_type_t map1[(int)GRID_WIDTH*GRID_HEIGHT];
+extern const cell_type_t map1[(int)GRID_WIDTH*(int)GRID_HEIGHT];
 
 #endif //MAPS_H
